lab04/nosso_propaganda/requirements.txt
RPi.GPIO
serial